+++ Sources/Certificate Pinning/TrustConfiguration.swift
@@ -9,7 +9,8 @@ import Foundation
 
 /** A configuration object which controls how a `TrustValidator` object makes decisions on which connections to accept and which to block. The configuration
     object consists of a series of configuration objects, each designed for a single domain. */
-@available(iOSApplicationExtension 10.0, tvOSApplicationExtension 10.0, watchOSApplicationExtension 3.0, *)
+@available(iOS, deprecated: 14.0, message: "Prefer the `NSPinnedDomains` Info.plist key")
+@available(macOS, deprecated: 11.0, message: "Prefer the `NSPinnedDomains` Info.plist key")
 public struct TrustConfiguration {
     
     /// The `CertificateExpirationPolicy` determines how the validation should proceed when the certificate being pinned has expired.
@@ -152,8 +153,8 @@ public struct TrustConfiguration {
 }
 
 // MARK: - TrustConfiguration conformance to ExpressibleByArrayLiteral
-
-@available(iOSApplicationExtension 10.0, tvOSApplicationExtension 10.0, watchOSApplicationExtension 3.0, *)
+@available(iOS, deprecated: 14.0, message: "Prefer the `NSPinnedDomains` Info.plist key")
+@available(macOS, deprecated: 11.0, message: "Prefer the `NSPinnedDomains` Info.plist key")
 extension TrustConfiguration: ExpressibleByArrayLiteral {
     public typealias ArrayLiteralElement = DomainConfiguration