+++ CHANGELOG.md
@@ -2,7 +2,9 @@
 
 #### Enhancements
 
-* None
+* Deprecate `TrustConfiguration`, `TrustValidator`, `CertificateHasher`, `AuthenticationChallenge` and `TrustValidatingTransportService` for iOS 14+ and macOS 11+ in favor of `NSPinnedDomains` Info.plist key
+[Will McGinty](https://github.com/wmcginty)
+[#154](https://github.com/BottleRocketStudios/iOS-Hyperspace/pull/154)
 
 ##### Bug Fixes